# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UK Football Fixed Odds Pool Betting API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      parameters:
        - { in: query, name: weekId, schema: { type: integer } }
        - { in: query, name: cashierId, schema: { type: integer } }
        - { in: query, name: agentId, schema: { type: integer } }
        - { in: query, name: status, schema: { type: string, enum: [active, cancelled, won, lost] } }
        - { in: query, name: scope, schema: { type: string, enum: [mine, shop, all] } }
      responses:
        "200":
          description: Tickets visible to current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTicketsResponse"
    post:
      operationId: createTicket
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTicketBody" }
      responses:
        "201":
          description: Ticket placed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "400":
          description: Validation failed
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
  /tickets/{id}/cancel:
    post:
      operationId: cancelTicket
      tags: [tickets]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ----- Auth -----
...
    Ticket:
      type: object
      properties:
        id: { type: integer }
        ticketCode: { type: string }
        weekId: { type: integer }
        weekNumber: { type: integer }
        cashierId: { type: integer }
        cashierName: { type: string }
        agentId: { type: integer }
        shopName: { type: string }
        betType: { type: string, enum: [nap, perm] }
        poolType: { type: string, enum: [single, double, nap, under3, under4, under5, under6] }
        oddsType: { type: string, enum: [standard, high] }
        oddsValue: { type: number }
        selectedNumbers:
          type: array
          items: { type: integer }
        totalLines: { type: integer }
        stake: { type: number }
        perLine: { type: number }
        status: { type: string, enum: [active, cancelled, won, lost] }
        winningLines: { type: integer }
        winnings: { type: number }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - ticketCode
        - weekId
        - weekNumber
        - cashierId
        - cashierName
        - agentId
        - shopName
        - betType
        - poolType
        - oddsType
        - oddsValue
        - selectedNumbers
        - totalLines
        - stake
        - perLine
        - status
        - winningLines
        - winnings
        - createdAt
    CreateTicketBody:
      type: object
      properties:
        weekId: { type: integer }
        betType: { type: string, enum: [nap, perm] }
        poolType: { type: string, enum: [single, double, nap, under3, under4, under5, under6] }
        oddsType: { type: string, enum: [standard, high] }
        selectedNumbers:
          type: array
          items: { type: integer }
        stake: { type: number }
      required: [weekId, betType, poolType, oddsType, selectedNumbers, stake]
    ListTicketsResponse:
      type: object
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - tickets
        - total
        - page
        - pageSize
        - totalPages